pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcriber"
version = "0.1.0"
description = "Local + Colab speech transcription with speaker diarization (WhisperX + pyannote)."
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = { file = "LICENSE" }
authors = [{ name = "Transcriber contributors" }]
keywords = ["whisper", "whisperx", "pyannote", "transcription", "diarization", "speech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "whisperx>=3.1.1",
    "typer[all]>=0.12",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
    "black>=24.0",
]

[project.scripts]
transcriber = "transcriber.cli:app"

[project.urls]
Homepage = "https://github.com/OWNER/REPO"
Issues = "https://github.com/OWNER/REPO/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/transcriber"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
